refactor: rename to trace reader
