+++ settings.gradle.kts
@@ -17,4 +17,4 @@ include(":hypertrace-trace-visualizer")
 include(":hypertrace-trace-enricher-api")
 include(":hypertrace-trace-enricher-impl")
 include(":hypertrace-trace-enricher")
-include(":trace-client")
+include(":trace-reader")
