+++ trace-reader/src/main/java/org/hypertrace/trace/reader/ValueResolver.java
@@ -1,4 +1,4 @@
-package org.hypertrace.trace.client;
+package org.hypertrace.trace.reader;
 
 import io.reactivex.rxjava3.core.Single;
 import org.hypertrace.core.attribute.service.cachingclient.CachingAttributeClient;
@@ -6,7 +6,7 @@ import org.hypertrace.core.attribute.service.projection.AttributeProjectionRegis
 import org.hypertrace.core.attribute.service.v1.AttributeMetadata;
 import org.hypertrace.core.attribute.service.v1.LiteralValue;
 
-public interface ValueResolver {
+interface ValueResolver {
 
   Single<LiteralValue> resolve(ValueSource valueSource, AttributeMetadata attributeMetadata);
