+++ trace-reader/src/main/java/org/hypertrace/trace/reader/ValueSource.java
@@ -1,4 +1,4 @@
-package org.hypertrace.trace.client;
+package org.hypertrace.trace.reader;
 
 import java.util.Optional;
 import org.hypertrace.core.attribute.service.v1.AttributeKind;
@@ -6,7 +6,7 @@ import org.hypertrace.core.attribute.service.v1.LiteralValue;
 import org.hypertrace.core.datamodel.Event;
 import org.hypertrace.core.datamodel.StructuredTrace;
 
-public interface ValueSource {
+interface ValueSource {
   Optional<LiteralValue> getAttribute(String key, AttributeKind attributeKind);
 
   Optional<LiteralValue> getMetric(String key, AttributeKind attributeKind);
